Move implementation to proper file
